fitbit-back/requirements.txt
fastapi[standard]
uvicorn[standard]
mysql-connector-python
requests
python-dotenv